requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorthief==0.2.1
discord.py==2.1.0
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-OAuthlib==0.9.6
Flask-WTF==1.0.1
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.2.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
oauthlib==2.1.0
Pillow==9.3.0
pymongo==4.3.3
python-dotenv==0.21.0
python-engineio==3.14.2
python-socketio==5.0.0
requests==2.28.1
requests-oauthlib==1.1.0
six==1.16.0
tornado==6.2
twitchio==2.5.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.2
zipp==3.11.0
